Fix EC_DOD_Light uninitialized fields and clean up compiler warnings
EC_DOD_Light had no default member initializers, so a scene's <Light>
entry omitting a field (e.g. <Intensity> or <Attenuation> on a
Point/Spot light) left that field as genuine garbage feeding directly
into computeLightCutoffRadius - not just an MSVC C4701 warning, a real
correctness bug. Every field now has a sensible default; the
attenuation default in particular routes through the fallback radius
computeLightCutoffRadius already has for the degenerate case.

Also addresses the rest of the project's /W4 warning list that was
worth acting on rather than just counting:
- sscanf -> TIXML_SSCANF (tinyxml's own portable macro) in
  EC_DOD_EntityFactory.cpp, removing the deprecated-CRT-call warnings.
- localtime -> localtime_s in ECX_Logging.cpp, removing both the
  warning and the shared-static-buffer thread-safety hazard.
- EC_ThreadManager::s_activeThreads changed int -> size_t to match
  what it's actually assigned from/used for, fixing a narrowing warning.
- A discarded [[nodiscard]] luabridge::push result in
  EC_LuaScriptingSystem.h (visible on every normal build, not just
  /W4) is now checked and logged on failure.
- Three initialized-but-unused locals: CubemapManager.cpp now checks
  and logs FBO completeness (matching ShadowAtlas.cpp/GBuffer.cpp's
  existing pattern) instead of discarding the status; TextureManager.cpp
  and EC_SceneManager.cpp had genuinely dead computations removed.

+++ src/TaskManager/EC_ThreadManager.cpp
@@ -29,7 +29,7 @@ std::deque<std::shared_ptr<EC_Task>> EC_ThreadManager::s_Tasks;
 std::mutex EC_ThreadManager::s_Lock;
 std::condition_variable EC_ThreadManager::s_Flag;
 bool EC_ThreadManager::s_running;
-int EC_ThreadManager::s_activeThreads;
+size_t EC_ThreadManager::s_activeThreads;
 
 EC_ThreadManager::EC_ThreadManager()
 {
